+++ pytorch_lightning/testing/lm_test_module_base.py
@@ -15,9 +15,10 @@ from pytorch_lightning.root_module.root_module import LightningModule
 from pytorch_lightning import data_loader
 
 
-class NoValModel(LightningModule):
+class LightningTestModelBase(LightningModule):
     """
-    Sample model to show how to define a template
+    Base LightningModule for testing. Implements only the required
+    interface
     """
 
     def __init__(self, hparams, force_remove_distributed_sampler=False):
@@ -26,7 +27,7 @@ class NoValModel(LightningModule):
         :param hparams:
         """
         # init superclass
-        super(NoValModel, self).__init__()
+        super(LightningTestModelBase, self).__init__()
         self.hparams = hparams
 
         self.batch_size = hparams.batch_size
@@ -109,9 +110,6 @@ class NoValModel(LightningModule):
         if self.trainer.batch_nb % 2 == 0:
             return loss_val
 
-    def on_tng_metrics(self, logs):
-        logs['some_tensor_to_test'] = torch.rand(1)
-
     # ---------------------
     # TRAINING SETUP
     # ---------------------
@@ -124,9 +122,9 @@ class NoValModel(LightningModule):
         optimizer = optim.Adam(self.parameters(), lr=self.hparams.learning_rate)
 
         # test returning only 1 list instead of 2
-        return [optimizer]
+        return optimizer
 
-    def __dataloader(self, train):
+    def _dataloader(self, train):
         # init data generators
         transform = transforms.Compose([transforms.ToTensor(),
                                         transforms.Normalize((0.5,), (1.0,))])
@@ -156,7 +154,7 @@ class NoValModel(LightningModule):
 
     @data_loader
     def tng_dataloader(self):
-        return self.__dataloader(train=True)
+        return self._dataloader(train=True)
 
     @staticmethod
     def add_model_specific_args(parent_parser, root_dir):  # pragma: no cover
