+++ tests/debug.py
@@ -1,6 +1,6 @@
 from pytorch_lightning import Trainer
 from examples import LightningTemplateModel
-from pytorch_lightning.testing import LightningTestModel, NoValEndTestModel, NoValModel
+from pytorch_lightning.testing import LightningTestModel
 from argparse import Namespace
 from test_tube import Experiment
 from pytorch_lightning.callbacks import ModelCheckpoint
